chore: update dependencies
Update the dependencies to their latest versions. This makes this
library requiring at least Rust 1.56.0.

Due to dependencies having updated the version of `bitvec` they are
using, the calls to it changed. In the previous version 0.22
[`as_buffer` returned] a reference to the underlying data. This
method was removed in version 1.0.0 in favour of making of making the
[`data` field public].

A newer version of nightly on CI is needed, due to the minimum supported
Rust version 1.56.0. Currently running on ARM via GitHub Actions doesn't
work due to invalid memory acccesses, hence use CircleCI to run on ARM.

[`as_buffer` returned]: https://docs.rs/bitvec/0.22.3/src/bitvec/array.rs.html#274-276
[`data` field public]: https://docs.rs/bitvec/1.0.0/src/bitvec/array.rs.html#33